Rename slackbot helper
