+++ lib/commands/subscribe.js
@@ -2,7 +2,7 @@ const {
   Subscribed, NotFound, AlreadySubscribed, NotSubscribed, UpdatedSettings,
 } = require('../messages/flow');
 
-const isSlackbot = require('../is-slackbot');
+const isRequestFromSlack = require('../is-request-from-slack');
 
 /**
  * Subscribes a slack channel to activity from an Organization or Repository
@@ -20,7 +20,7 @@ module.exports = async (req, res) => {
   req.log.debug({ installation, resource }, 'Lookup respository to subscribe');
 
   async function respondWith(message) {
-    if (!isSlackbot(req)) {
+    if (!isRequestFromSlack(req)) {
       res.redirect(`https://slack.com/app_redirect?channel=${command.channel_id}&team=${command.team_id}`);
     }
