+++ lib/error-handler.js
@@ -2,7 +2,7 @@ const Raven = require('raven');
 const timeout = require('connect-timeout');
 
 const { Exception } = require('./messages/flow');
-const isSlackbot = require('./is-slackbot');
+const isRequestFromSlack = require('./is-request-from-slack');
 
 module.exports = {
   setup: (app) => {
@@ -46,7 +46,7 @@ module.exports = {
         return command.respond(message);
       } else if (res.headersSent) {
         req.log.warn({ err }, 'Headers were already sent, so swallowing error');
-      } else if (isSlackbot(req)) {
+      } else if (isRequestFromSlack(req)) {
         res.json(message);
       } else {
         res.status(500).send(`An unexpected error occured: ${res.sentry}`);
