+++ lib/is-request-from-slack.js
