+++ lib/middleware/get-installation.js
@@ -1,6 +1,6 @@
 const { InstallGitHubApp, NotFound } = require('../messages/flow');
 const SignedParams = require('../signed-params');
-const isSlackbot = require('../is-slackbot');
+const isRequestFromSlack = require('../is-request-from-slack');
 
 /**
  * Get the installation for the given account name
@@ -40,12 +40,12 @@ module.exports = async function getInstallation(req, res, next) {
 
     const installLink = `${protocol}://${host}/github/install/${owner.id}/${await state.stringify()}`;
 
-    if (isSlackbot(req)) {
+    if (isRequestFromSlack(req)) {
       return command.respond(new InstallGitHubApp({ installLink, owner, gitHubUser }));
     }
     return res.redirect(installLink);
   } catch (err) {
-    if (!isSlackbot(req)) {
+    if (!isRequestFromSlack(req)) {
       res.redirect(`https://slack.com/app_redirect?channel=${command.channel_id}&team=${command.team_id}`);
     }
     return command.respond(new NotFound(command.args[0]));
